반복문 연습

+++ SunJae_GameCoding_Practice/SunJae_GameCoding_Practice/SunJae_GameCoding_Practice.cpp
@@ -82,7 +82,7 @@ int main()
 
 	// 반복문 : for , while
 
-	for (size_t i = 0; i < 3; i++)
+	for (int i = 0; i < 3; i++)
 	{
 
 	}
